Create a userSettings Utility
