+++ lib/ui/utils/_utils.mjs
@@ -3,11 +3,14 @@
 
 The MAPP UI utils exports legacy methods to warn that plugins are required for the use of third party dataview libraries eg. Tabulator and Chart[js].
 
-The idleLogout, imagePreview, resizeHandler utility methods are imported and added to the ui.utils{} object.
+The idleLogout, imagePreview, resizeHandler, jsonDataview, locationCount and userSettings utility methods are imported and added to the ui.utils{} object.
 
 @requires /ui/utils/idleLogout
 @requires /ui/utils/imagePreview
 @requires /ui/utils/resizeHandler
+@requires /ui/utils/jsonDataview
+@requires /ui/utils/locationCount
+@requires /ui/utils/userSettings
 
 @module /ui/utils
 */
@@ -52,6 +55,8 @@ import jsonDataview from './jsonDataview.mjs'
 
 import locationCount from './locationCount.mjs'
 
+import userSettings from './userSettings.mjs'
+
 export default {
   Chart,
   Tabulator,
@@ -59,5 +64,6 @@ export default {
   idleLogout,
   imagePreview,
   resizeHandler,
-  locationCount
+  locationCount,
+  userSettings
 }
